refactor: dir structure
