+++ src/app/(pages)/article/[id]/page.jsx
