+++ src/app/(pages)/layout.jsx
@@ -1,4 +1,3 @@
-import "@/styles/style.scss";
 import Header from "@/components/layouts/Header";
 import Footer from "@/components/layouts/Footer";
 
@@ -9,10 +8,10 @@ export const metadata = {
 
 export default function RootLayout({ children }) {
   return (
-    <span>
+    <>
       <Header />
       {children}
       <Footer />
-    </span>
+    </>
   );
 }
