+++ src/app/(pages)/register/page.jsx
